+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Pending
 
 * Add `uninstall_file_onerror_exit` to API
+* Enable Windows long path support for `setup.exe` and `dkml-package-entry.exe` in application manifest. Confer <https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation>
 
 ## 0.5.2